crain_interface/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.3.0
httpx>=0.25.0
jinja2>=3.1.2
python-dotenv>=1.0.0
python-multipart>=0.0.6